+++ templates/withKibana/transactions.go
@@ -74,7 +74,7 @@ var Transactions = Object{
 				"type": "keyword",
 			},
 			"nonce": Object{
-				"type": "long",
+				"type": "double",
 			},
 			"operation": Object{
 				"type": "keyword",
@@ -92,7 +92,7 @@ var Transactions = Object{
 				"type": "long",
 			},
 			"round": Object{
-				"type": "long",
+				"type": "double",
 			},
 			"searchOrder": Object{
 				"type": "long",
